refactor: move types.py to statustypes.py to avoid confusion with stdlib
There's a (somewhat rarely-used) stdlib module named "types", so it
seems best to avoid confusion with this.

Also updates the dev dependencies (it was the upgrade of ruff that
complained about this).
