+++ pyproject.toml
@@ -105,7 +105,7 @@ convention = "google"
 builtins-ignorelist = ["id", "min", "map", "range", "type", "input", "format"]
 
 [tool.ruff.lint.per-file-ignores]
-"jubilant/types.py" = [
+"jubilant/statustypes.py" = [
     # Missing docstring in public class
     "D101",
 ]
